RGW - Split rgw_user into generic and radosstore
rgw_user.[cc,h] contained parts that were generic, and parts that were
for radosstore.  Split them into common code (rgw_user.[cc,h]) and
specific code (rgw_user_ctl.[cc,h]).
