+++ src/rgw/driver/rados/rgw_object_expirer_core.cc
@@ -22,7 +22,7 @@
 #include "include/utime.h"
 #include "include/str_list.h"
 
-#include "rgw_user.h"
+#include "rgw_user_ctl.h"
 #include "rgw_bucket.h"
 #include "rgw_acl.h"
 #include "rgw_acl_s3.h"
