+++ src/rgw/driver/rados/rgw_putobj_processor.h
@@ -19,6 +19,7 @@
 
 #include "rgw_putobj.h"
 #include "services/svc_tier_rados.h"
+#include "rgw_tools.h"
 #include "rgw_sal.h"
 #include "rgw_obj_manifest.h"
