+++ src/rgw/driver/rados/rgw_sal_rados.cc
@@ -67,6 +67,7 @@
 #include "rgw_zone.h"
 #include "rgw_restore.h"
 #include "rgw_multipart_meta_filter.h"
+#include "rgw_user_ctl.h"
 
 #include "services/svc_bilog_rados.h"
 #include "services/svc_bi_rados.h"
