+++ src/rgw/driver/rados/rgw_sync_error_repo.cc
@@ -14,6 +14,7 @@
 
 #include "rgw_sync_error_repo.h"
 #include "rgw_coroutine.h"
+#include "rgw_tools.h"
 #include "rgw_sal.h"
 #include "cls/cmpomap/client.h"
