+++ src/rgw/driver/rados/roles.cc
@@ -20,6 +20,7 @@
 #include "common/dout.h"
 #include "cls/user/cls_user_client.h"
 #include "rgw_role.h"
+#include "rgw_tools.h"
 #include "rgw_sal.h"
 
 namespace rgwrados::roles {
