+++ src/rgw/driver/rados/topics.cc
@@ -19,6 +19,7 @@
 #include "common/dout.h"
 #include "cls/user/cls_user_client.h"
 #include "rgw_pubsub.h"
+#include "rgw_tools.h"
 #include "rgw_sal.h"
 
 namespace rgwrados::topics {
