+++ src/rgw/driver/rados/users.cc
@@ -20,7 +20,7 @@
 #include "common/dout.h"
 #include "cls/user/cls_user_client.h"
 #include "rgw_common.h"
-#include "rgw_sal.h"
+#include "rgw_tools.h"
 
 namespace rgwrados::users {
