+++ src/rgw/rgw_arn.h
@@ -7,8 +7,8 @@
 
 #include <fmt/ostream.h>
 
-class rgw_obj;
-class rgw_bucket;
+struct rgw_obj;
+struct rgw_bucket;
 
 namespace rgw {
