+++ src/rgw/rgw_file_int.h
@@ -29,7 +29,6 @@
 #include "common/cohort_lru.h"
 #include "common/ceph_timer.h"
 #include "rgw_common.h"
-#include "driver/rados/rgw_user.h"
 #include "rgw_lib.h"
 #include "rgw_ldap.h"
 #include "rgw_token.h"
