+++ src/rgw/rgw_lc_s3.cc
@@ -1,14 +1,11 @@
 // -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:nil -*-
 // vim: ts=8 sw=2 sts=2 expandtab ft=cpp
 
-#include <string.h>
-
 #include <iostream>
 #include <map>
 
 #include "include/types.h"
 
-#include "driver/rados/rgw_user.h"
 #include "rgw_lc_s3.h"
