+++ src/rgw/rgw_request.h
@@ -5,7 +5,6 @@
 
 #include "rgw_common.h"
 #include "rgw_acl.h"
-#include "driver/rados/rgw_user.h"
 #include "rgw_op.h"
 
 #include "common/QueueRing.h"
