+++ src/rgw/rgw_sts.cc
@@ -22,7 +22,6 @@
 #include "rgw_b64.h"
 #include "rgw_common.h"
 #include "rgw_role.h"
-#include "driver/rados/rgw_user.h"
 #include "rgw_iam_policy.h"
 #include "rgw_sts.h"
 #include "rgw_sal.h"
