+++ src/rgw/rgw_user.h
@@ -5,101 +5,14 @@
 
 #include <string>
 #include <boost/algorithm/string.hpp>
-#include "include/ceph_assert.h"
 
-#include "include/types.h"
 #include "rgw_common.h"
+#include "rgw_formats.h"
 #include "rgw_tools.h"
 
-#include "rgw_string.h"
-
-#include "common/Formatter.h"
-#include "rgw_formats.h"
 #include "rgw_sal_fwd.h"
 
-#define RGW_USER_ANON_ID "anonymous"
-
-constexpr auto SECRET_KEY_LEN=40;
-constexpr auto PUBLIC_ID_LEN=20;
-constexpr auto RAND_SUBUSER_LEN=5;
-
 constexpr auto XMLNS_AWS_S3 = "http://s3.amazonaws.com/doc/2006-03-01/";
-
-class RGWUserCtl;
-class RGWBucketCtl;
-class RGWUserBuckets;
-class RGWMetadataHandler;
-class RGWSI_User;
-
-// generate a random secret access key of SECRET_KEY_LEN=40
-void rgw_generate_secret_key(CephContext* cct,
-                             std::string& secret_key);
-
-// generate a unique random access key id of PUBLIC_ID_LEN=20
-int rgw_generate_access_key(const DoutPrefixProvider* dpp,
-                            optional_yield y,
-                            rgw::sal::Driver* driver,
-                            std::string& access_key_id);
-
-/**
- * A string wrapper that includes encode/decode functions for easily accessing
- * a UID in all forms. In some objects, this may refer to an account id instead
- * of a user.
- */
-struct RGWUID
-{
-  std::string id;
-  void encode(bufferlist& bl) const {
-    using ceph::encode;
-    encode(id, bl);
-  }
-  void decode(bufferlist::const_iterator& bl) {
-    using ceph::decode;
-    decode(id, bl);
-  }
-  void dump(Formatter *f) const {
-    f->dump_string("user_id", id);
-  }
-  static std::list<RGWUID> generate_test_instances() {
-    std::list<RGWUID> o;
-    o.emplace_back();
-    o.emplace_back();
-    o.back().id = "test:tester";
-    return o;
-  }
-};
-WRITE_CLASS_ENCODER(RGWUID)
-
-/** Entry for bucket metadata collection */
-struct bucket_meta_entry {
-  size_t size;
-  size_t size_rounded;
-  ceph::real_time creation_time;
-  uint64_t count;
-};
-
-int rgw_sync_all_stats(const DoutPrefixProvider *dpp,
-                       optional_yield y, rgw::sal::Driver* driver,
-                       const rgw_owner& owner, const std::string& tenant);
-extern int rgw_user_get_all_buckets_stats(const DoutPrefixProvider *dpp,
-  rgw::sal::Driver* driver, rgw::sal::User* user,
-  std::map<std::string, bucket_meta_entry>& buckets_usage_map, optional_yield y);
-
-/**
- * Get the anonymous (ie, unauthenticated) user info.
- */
-extern void rgw_get_anon_user(RGWUserInfo& info);
-
-extern void rgw_perm_to_str(uint32_t mask, char *buf, int len);
-extern uint32_t rgw_str_to_perm(const char *str);
-
-extern int rgw_validate_tenant_name(const std::string& t);
-
-enum ObjectKeyType {
-  KEY_TYPE_SWIFT,
-  KEY_TYPE_S3,
-  KEY_TYPE_UNDEFINED
-};
 
 enum RGWKeyPoolOp {
   GENERATE_KEY,
@@ -112,6 +25,26 @@ enum RGWUserId {
   RGW_USER_EMAIL,
   RGW_ACCESS_KEY,
 };
+
+enum ObjectKeyType {
+  KEY_TYPE_SWIFT,
+  KEY_TYPE_S3,
+  KEY_TYPE_UNDEFINED
+};
+
+constexpr auto SECRET_KEY_LEN=40;
+constexpr auto PUBLIC_ID_LEN=20;
+constexpr auto RAND_SUBUSER_LEN=5;
+
+// generate a random secret access key of SECRET_KEY_LEN=40
+void rgw_generate_secret_key(CephContext* cct,
+                             std::string& secret_key);
+
+// generate a unique random access key id of PUBLIC_ID_LEN=20
+int rgw_generate_access_key(const DoutPrefixProvider* dpp,
+                            optional_yield y,
+                            rgw::sal::Driver* driver,
+                            std::string& access_key_id);
 
 /*
  * An RGWUser class along with supporting classes created
@@ -469,6 +402,78 @@ struct RGWUserAdminOpState {
   RGWUserAdminOpState(rgw::sal::Driver* driver);
 };
 
+/* Wrappers for admin API functionality */
+
+class RGWUserAdminOp_User
+{
+public:
+  static int list(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
+                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher);
+
+  static int info(const DoutPrefixProvider *dpp,
+		  rgw::sal::Driver* driver,
+                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		  bool dump_keys, optional_yield y);
+
+  static int create(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+
+  static int modify(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher, optional_yield y);
+
+  static int remove(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
+                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher, optional_yield y);
+};
+
+class RGWUserAdminOp_Subuser
+{
+public:
+  static int create(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+
+  static int modify(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+
+  static int remove(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+};
+
+class RGWUserAdminOp_Key
+{
+public:
+  static int create(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+
+  static int remove(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+};
+
+class RGWUserAdminOp_Caps
+{
+public:
+  static int add(const DoutPrefixProvider *dpp,
+		 rgw::sal::Driver* driver,
+		 RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		 optional_yield y);
+
+  static int remove(const DoutPrefixProvider *dpp,
+		    rgw::sal::Driver* driver,
+		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
+		    optional_yield y);
+};
+
 class RGWUser;
 
 class RGWAccessKeyPool
@@ -659,202 +664,28 @@ public:
   friend class RGWUserCapPool;
 };
 
-/* Wrappers for admin API functionality */
-
-class RGWUserAdminOp_User
-{
-public:
-  static int list(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
-                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher);
-
-  static int info(const DoutPrefixProvider *dpp,
-		  rgw::sal::Driver* driver,
-                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		  bool dump_keys, optional_yield y);
-
-  static int create(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
-
-  static int modify(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher, optional_yield y);
-
-  static int remove(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
-                  RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher, optional_yield y);
+/** Entry for bucket metadata collection */
+struct bucket_meta_entry {
+  size_t size;
+  size_t size_rounded;
+  ceph::real_time creation_time;
+  uint64_t count;
 };
 
-class RGWUserAdminOp_Subuser
-{
-public:
-  static int create(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
+int rgw_sync_all_stats(const DoutPrefixProvider *dpp,
+                       optional_yield y, rgw::sal::Driver* driver,
+                       const rgw_owner& owner, const std::string& tenant);
+extern int rgw_user_get_all_buckets_stats(const DoutPrefixProvider *dpp,
+  rgw::sal::Driver* driver, rgw::sal::User* user,
+  std::map<std::string, bucket_meta_entry>& buckets_usage_map, optional_yield y);
 
-  static int modify(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
+/**
+ * Get the anonymous (ie, unauthenticated) user info.
+ */
+extern void rgw_get_anon_user(RGWUserInfo& info);
 
-  static int remove(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
-};
+extern void rgw_perm_to_str(uint32_t mask, char *buf, int len);
+extern uint32_t rgw_str_to_perm(const char *str);
 
-class RGWUserAdminOp_Key
-{
-public:
-  static int create(const DoutPrefixProvider *dpp, rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
+extern int rgw_validate_tenant_name(const std::string& t);
 
-  static int remove(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
-};
-
-class RGWUserAdminOp_Caps
-{
-public:
-  static int add(const DoutPrefixProvider *dpp,
-		 rgw::sal::Driver* driver,
-		 RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		 optional_yield y);
-
-  static int remove(const DoutPrefixProvider *dpp,
-		    rgw::sal::Driver* driver,
-		    RGWUserAdminOpState& op_state, RGWFormatterFlusher& flusher,
-		    optional_yield y);
-};
-
-class RGWUserCtl
-{
-  struct Svc {
-    RGWSI_Zone *zone{nullptr};
-    RGWSI_User *user{nullptr};
-  } svc;
-
-  struct Ctl {
-    RGWBucketCtl *bucket{nullptr};
-  } ctl;
-
-public:
-  RGWUserCtl(RGWSI_Zone *zone_svc, RGWSI_User *user_svc);
-
-  void init(RGWBucketCtl *bucket_ctl) {
-    ctl.bucket = bucket_ctl;
-  }
-
-  RGWBucketCtl *get_bucket_ctl() {
-    return ctl.bucket;
-  }
-
-  struct GetParams {
-    RGWObjVersionTracker *objv_tracker{nullptr};
-    ceph::real_time *mtime{nullptr};
-    rgw_cache_entry_info *cache_info{nullptr};
-    std::map<std::string, bufferlist> *attrs{nullptr};
-
-    GetParams() {}
-
-    GetParams& set_objv_tracker(RGWObjVersionTracker *_objv_tracker) {
-      objv_tracker = _objv_tracker;
-      return *this;
-    }
-
-    GetParams& set_mtime(ceph::real_time *_mtime) {
-      mtime = _mtime;
-      return *this;
-    }
-
-    GetParams& set_cache_info(rgw_cache_entry_info *_cache_info) {
-      cache_info = _cache_info;
-      return *this;
-    }
-
-    GetParams& set_attrs(std::map<std::string, bufferlist> *_attrs) {
-      attrs = _attrs;
-      return *this;
-    }
-  };
-
-  struct PutParams {
-    RGWUserInfo *old_info{nullptr};
-    RGWObjVersionTracker *objv_tracker{nullptr};
-    ceph::real_time mtime;
-    bool exclusive{false};
-    std::map<std::string, bufferlist> *attrs{nullptr};
-
-    PutParams() {}
-
-    PutParams& set_old_info(RGWUserInfo *_info) {
-      old_info = _info;
-      return *this;
-    }
-
-    PutParams& set_objv_tracker(RGWObjVersionTracker *_objv_tracker) {
-      objv_tracker = _objv_tracker;
-      return *this;
-    }
-
-    PutParams& set_mtime(const ceph::real_time& _mtime) {
-      mtime = _mtime;
-      return *this;
-    }
-
-    PutParams& set_exclusive(bool _exclusive) {
-      exclusive = _exclusive;
-      return *this;
-    }
-
-    PutParams& set_attrs(std::map<std::string, bufferlist> *_attrs) {
-      attrs = _attrs;
-      return *this;
-    }
-  };
-
-  struct RemoveParams {
-    RGWObjVersionTracker *objv_tracker{nullptr};
-
-    RemoveParams() {}
-
-    RemoveParams& set_objv_tracker(RGWObjVersionTracker *_objv_tracker) {
-      objv_tracker = _objv_tracker;
-      return *this;
-    }
-  };
-
-  int get_info_by_uid(const DoutPrefixProvider *dpp, 
-                      const rgw_user& uid, RGWUserInfo *info,
-                      optional_yield y, const GetParams& params = {});
-  int get_info_by_email(const DoutPrefixProvider *dpp, 
-                        const std::string& email, RGWUserInfo *info,
-                        optional_yield y, const GetParams& params = {});
-  int get_info_by_swift(const DoutPrefixProvider *dpp, 
-                        const std::string& swift_name, RGWUserInfo *info,
-                        optional_yield y, const GetParams& params = {});
-  int get_info_by_access_key(const DoutPrefixProvider *dpp, 
-                             const std::string& access_key, RGWUserInfo *info,
-                             optional_yield y, const GetParams& params = {});
-
-  int get_attrs_by_uid(const DoutPrefixProvider *dpp, 
-                       const rgw_user& user_id,
-                       std::map<std::string, bufferlist> *attrs,
-                       optional_yield y,
-                       RGWObjVersionTracker *objv_tracker = nullptr);
-
-  int store_info(const DoutPrefixProvider *dpp, 
-                 const RGWUserInfo& info, optional_yield y,
-                 const PutParams& params = {});
-  int remove_info(const DoutPrefixProvider *dpp, 
-                  const RGWUserInfo& info, optional_yield y,
-                  const RemoveParams& params = {});
-};
-
-// user metadata handler factory
-auto create_user_metadata_handler(RGWSI_User *user_svc)
-    -> std::unique_ptr<RGWMetadataHandler>;
